# kratos/api/explorer/v1/common.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.6
// 	protoc        v5.29.3
// source: explorer/v1/common.proto

package v1

import (
	_ "google.golang.org/genproto/googleapis/api/annotations"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type Status struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Code          int32                  `protobuf:"varint,1,opt,name=code,proto3" json:"code,omitempty"`
	Message       string                 `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *Status) Reset() {
	*x = Status{}
	mi := &file_explorer_v1_common_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *Status) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*Status) ProtoMessage() {}

func (x *Status) ProtoReflect() protoreflect.Message {
	mi := &file_explorer_v1_common_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use Status.ProtoReflect.Descriptor instead.
func (*Status) Descriptor() ([]byte, []int) {
	return file_explorer_v1_common_proto_rawDescGZIP(), []int{0}
}

func (x *Status) GetCode() int32 {
	if x != nil {
		return x.Code
	}
	return 0
}

func (x *Status) GetMessage() string {
	if x != nil {
		return x.Message
	}
	return ""
}

type PingReply struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Status        *Status                `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *PingReply) Reset() {
	*x = PingReply{}
	mi := &file_explorer_v1_common_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *PingReply) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*PingReply) ProtoMessage() {}

func (x *PingReply) ProtoReflect() protoreflect.Message {
	mi := &file_explorer_v1_common_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use PingReply.ProtoReflect.Descriptor instead.
func (*PingReply) Descriptor() ([]byte, []int) {
	return file_explorer_v1_common_proto_rawDescGZIP(), []int{1}
}

func (x *PingReply) GetStatus() *Status {
	if x != nil {
		return x.Status
	}
	return nil
}

var File_explorer_v1_common_proto protoreflect.FileDescriptor

const file_explorer_v1_common_proto_rawDesc = "" +
	"\n" +
	"\x18explorer/v1/common.proto\x12\x0fapi.explorer.v1\x1a\x1bgoogle/protobuf/empty.proto\x1a\x1cgoogle/api/annotations.proto\"6\n" +
	"\x06Status\x12\x12\n" +
	"\x04code\x18\x01 \x01(\x05R\x04code\x12\x18\n" +
	"\amessage\x18\x02 \x01(\tR\amessage\"<\n" +
	"\tPingReply\x12/\n" +
	"\x06status\x18\x01 \x01(\v2\x17.api.explorer.v1.StatusR\x06status2X\n" +
	"\x05Basic\x12O\n" +
	"\x04Ping\x12\x16.google.protobuf.Empty\x1a\x1a.api.explorer.v1.PingReply\"\x13\x82\xd3\xe4\x93\x02\r\x12\v/basic/pingB0\n" +
	"\x0fapi.explorer.v1P\x01Z\x1bexplorer/api/explorer/v1;v1b\x06proto3"

var (
	file_explorer_v1_common_proto_rawDescOnce sync.Once
	file_explorer_v1_common_proto_rawDescData []byte
)

func file_explorer_v1_common_proto_rawDescGZIP() []byte {
	file_explorer_v1_common_proto_rawDescOnce.Do(func() {
		file_explorer_v1_common_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_explorer_v1_common_proto_rawDesc), len(file_explorer_v1_common_proto_rawDesc)))
	})
	return file_explorer_v1_common_proto_rawDescData
}

var file_explorer_v1_common_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_explorer_v1_common_proto_goTypes = []any{
	(*Status)(nil),        // 0: api.explorer.v1.Status
	(*PingReply)(nil),     // 1: api.explorer.v1.PingReply
	(*emptypb.Empty)(nil), // 2: google.protobuf.Empty
}
var file_explorer_v1_common_proto_depIdxs = []int32{
	0, // 0: api.explorer.v1.PingReply.status:type_name -> api.explorer.v1.Status
	2, // 1: api.explorer.v1.Basic.Ping:input_type -> google.protobuf.Empty
	1, // 2: api.explorer.v1.Basic.Ping:output_type -> api.explorer.v1.PingReply
	2, // [2:3] is the sub-list for method output_type
	1, // [1:2] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_explorer_v1_common_proto_init() }
func file_explorer_v1_common_proto_init() {
	if File_explorer_v1_common_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_explorer_v1_common_proto_rawDesc), len(file_explorer_v1_common_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_explorer_v1_common_proto_goTypes,
		DependencyIndexes: file_explorer_v1_common_proto_depIdxs,
		MessageInfos:      file_explorer_v1_common_proto_msgTypes,
	}.Build()
	File_explorer_v1_common_proto = out.File
	file_explorer_v1_common_proto_goTypes = nil
	file_explorer_v1_common_proto_depIdxs = nil
}